+++ apps/api/plane/db/models/user.py
@@ -53,6 +53,11 @@ class BotTypeEnum(models.TextChoices):
     WORKSPACE_SEED = "WORKSPACE_SEED", "Workspace Seed"
 
 
+class TimestampDisplay(models.TextChoices):
+    EXACT = "exact", "Exact"
+    RELATIVE = "relative", "Relative"
+
+
 class User(AbstractBaseUser, PermissionsMixin):
     id = models.UUIDField(default=uuid.uuid4, unique=True, editable=False, db_index=True, primary_key=True)
     username = models.CharField(max_length=128, unique=True)
@@ -249,7 +254,7 @@ class Profile(TimeAuditModel):
     mobile_timezone_auto_set = models.BooleanField(default=False)
     # language
     language = models.CharField(max_length=255, default="en")
-    timestamp_display = models.CharField(max_length=16, default="exact")
+    timestamp_display = models.CharField(max_length=16, choices=TimestampDisplay.choices, default=TimestampDisplay.EXACT)
     start_of_the_week = models.PositiveSmallIntegerField(choices=START_OF_THE_WEEK_CHOICES, default=SUNDAY)
     goals = models.JSONField(default=dict)
     background_color = models.CharField(max_length=255, default=get_random_color)